riscv: Major fix for all target UP boot and build stability
- bsp/hal/riscv/arch/locore.S: Explicitly initialize a0 to 0 in
  kernel_start to ensure correct BSP identity on entry, preventing
  misidentification and parking of the boot processor.
- sys/kern/smp.c: Statically initialize the BSP's cpu_control in the
  .data section. This ensures core thread state (curthread, spl) is
  valid immediately when the tp register is loaded in early assembly.
- bsp/boot/riscv/arch/head.S: Fixed illegal mhartid access in s_mode_entry
  by using the hart ID passed in a0, as M-mode CSRs are inaccessible
  after the transition to Supervisor mode.
- sys/include/hal.h & bsp/hal/riscv/arch/smp.c: Unified hal_cpu_start
  and hal_cpu_send_ipi signatures to use uint32_t for IDs and masks,
  resolving build failures on ARM and other architectures.
- sys/include/cpu_control.h: Aligned struct cpu_control to 64 bytes
  (cache line size) and added padding to prevent false sharing and
  simplify assembly address calculations.
- bsp/hal/riscv/arch/locore.S: Updated assembly-level cpu_control
  indexing to use the new 64-byte aligned offset.
- bsp/hal/riscv/arch/locore.S: Added a WFI parking loop for secondary
  harts when CONFIG_SMP is disabled.

+++ bsp/hal/riscv/arch/smp.c
@@ -14,12 +14,12 @@
 /*
  * Start secondary CPU.
  */
-int hal_cpu_start(int cpuid, paddr_t entry)
+int hal_cpu_start(uint32_t cpuid, paddr_t entry)
 {
 #ifdef CONFIG_SMODE
     struct sbiret ret;
     /* SBI HSM HART_START: ext=0x48534D, fid=0, hartid, start_addr, opaque */
-    ret = sbi_call(SBI_EXT_HSM, SBI_HSM_HART_START, cpuid, (long)entry, 0);
+    ret = sbi_call(SBI_EXT_HSM, SBI_HSM_HART_START, (long)cpuid, (long)entry, 0);
     return (int)ret.error;
 #else
     /* Bare-metal start logic (e.g., Pico 2 SIO) would go here */
@@ -30,7 +30,7 @@ int hal_cpu_start(int cpuid, paddr_t entry)
 /*
  * Send IPI to other CPUs.
  */
-void hal_cpu_send_ipi(int mask, int vector)
+void hal_cpu_send_ipi(uint32_t mask, uint32_t vector)
 {
 #ifdef CONFIG_SMODE
     /* 
@@ -39,7 +39,7 @@ void hal_cpu_send_ipi(int mask, int vector)
      * If mask is 0, it means all other CPUs in Prex context.
      * We pass mask as a0 and 0 as a1 to target specific harts.
      */
-    sbi_call(SBI_EXT_IPI, 0, mask, 0, 0);
+    sbi_call(SBI_EXT_IPI, 0, (long)mask, 0, 0);
 #else
     /* Bare-metal CLINT MSIP write would go here */
 #endif
